+++ BlishHud-Raid-Clears/Settings/Models/RaidSettings.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Blish_HUD.Settings;
+using Microsoft.Xna.Framework;
 using RaidClears.Features.Raids.Services;
 
 namespace RaidClears.Settings.Models;
@@ -15,6 +16,8 @@ public class RaidSettings
     public SettingEntry<bool> RaidPanelHighlightCotM { get; set; }
     public SettingEntry<bool> RaidPanelMentorProgress { get; set; }
     public SettingEntry<bool> RaidPanelMentorProgressPopup { get; set; }
+    public SettingEntry<bool> RaidPanelMentorProgressPopupReposition { get; set; }
+    public SettingEntry<Point> RaidPanelMentorProgressPopupPosition { get; set; }
     public SettingEntry<string> RaidPanelColorEmbolden { get; set; }
     public SettingEntry<string> RaidPanelColorCotm { get; set; }
     public SettingEntry<string> RaidPanelColorNonWeeklyBounty { get; set; }
@@ -42,6 +45,8 @@ public class RaidSettings
         RaidPanelOmitEventEncounters = settings.DefineSetting(Settings.Raids.Module.omitEventEncounters);
         RaidPanelMentorProgress = settings.DefineSetting(Settings.Raids.Module.mentorProgress);
         RaidPanelMentorProgressPopup = settings.DefineSetting(Settings.Raids.Module.mentorProgressPopup);
+        RaidPanelMentorProgressPopupReposition = settings.DefineSetting(Settings.Raids.Module.mentorProgressPopupReposition);
+        RaidPanelMentorProgressPopupPosition = settings.DefineSetting(Settings.Raids.Module.mentorProgressPopupPosition);
         
         Style = new DisplayStyle
         {